Master (#90)
* Release 4.4 (#84)

* Master in dev (#68)

* New release (#67)

* Release :: 3.5.0 (#59)

* Feature ::retrieve output domain thanks to pagination (#62)

* Release :: 3.6.0

* useless file

* Feature :: a real math postprocess (#66)

* Feature :: a real math postprocess :tada:

* better message

* Better signature for converters (same as others) (#65)

* Feature :: text postprocess (#52)

* Init

* Push bad import

* expose string methods

* nice doc

* improved

* New version

* fix typo

* Feature :: add traceback + clean notebooks (#70)

* multiple updates for sdk

* to simplify test

* remove x for pytest

* bum version sdk

* Feature :: add lib scipy (#71)

* add change_date_format function (#73)

* add change_date_format function

* correct test coverage

* correct function signature

* switch to new_column

* flake8 :: new errors detected with upgrade

* Simplify function

* Switch param dst_column to new_column (#74)

* switch formula drom dst_column to new_column

* switch replace into new_column

* clean test_formula and test_replace fonctions

* transfer handle_deprecated_new_colum into helpers

* bump version and remove decorator (in laputa now)

* flake8

* new_column for test postprocess functions

* add-argmin (#76)

* Example :: add import getpass (#78)

* explicit sort=False (#80)

* explicit sort=Fals

* vincent's remarks

* f/Uniformise new_column behavior (#77)

* init

* uniformisation of new_column behavior

* add test

* tipos

* simplify test and fix convert_datetime_to_str

* Hotfix on formula (#79)

* sauv

* add _ in isword function

* Allow - in column names for formula

* improve formula

* remove useless loggers

* fix merge

* add new utc param toucan is going international

* tipos

* f/change name of param and behvior enhance (#83)

* change name of param and behvior enhance

* small fixes

* Feature :: drop duplicates postprocess (#89)
