+++ toucan_data_sdk/utils/postprocess/__init__.py
@@ -2,12 +2,12 @@ from .argmax import argmax, argmin
 from .converter import convert_str_to_datetime, convert_datetime_to_str, cast, change_date_format
 from .cumsum import cumsum
 from .fillna import fillna
+from .filter import drop_duplicates, query, query_df
 from .groupby import groupby
 from .math import multiply, divide, add, subtract, formula
 from .melt import melt
 from .percentage import percentage
 from .pivot import pivot, pivot_by_group
-from .query_df import query, query_df
 from .rank import rank
 from .rename import rename
 from .replace import replace
